+++ packages/n8n-node/nodes/Probo/actions/organization/getAll.operation.ts
@@ -70,10 +70,6 @@ export async function execute(
 							organization {
 								id
 								name
-								description
-								websiteUrl
-								email
-								headquarterAddress
 								logo {
 									id
 									fileName
@@ -125,4 +121,3 @@ export async function execute(
 		pairedItem: { item: itemIndex },
 	};
 }
-